Solucion a un error con los tipos del jflex

+++ Proyecto1-Compi/src/ParserLexer/BasicLexerCupV.java
@@ -1163,7 +1163,7 @@ private Symbol symbol(int type, Object value) {
           case 6:
             { try {
             if (yytext().contains(".")) {
-                return symbol(sym.L_FLOAT, Double.parseDouble(yytext())); // Si es decimal
+                return symbol(sym.L_FLOAT, Float.parseFloat(yytext())); // Si es decimal
             } else {
                 return symbol(sym.L_INTEGER, Integer.parseInt(yytext())); // Si es entero
             }
